Login and Register Validate

+++ src/pages/Register.jsx
@@ -5,15 +5,13 @@ import { eye } from "react-icons-kit/icomoon/eye";
 import { eyeBlocked } from "react-icons-kit/icomoon/eyeBlocked";
 import Icon from "react-icons-kit";
 import { Link, useNavigate } from "react-router-dom";
-import "../pages/Register.css"
-
+import "../pages/Register.css";
 
 let ls = localStorage.getItem("Users");
 let table = ls ? JSON.parse(ls) : [];
 
 function Register() {
-
-  let nav=useNavigate();
+  let nav = useNavigate();
   const [types, setTypes] = useState("password");
   const [types1, setTypes1] = useState("password");
   const [icon, setIcon] = useState(eyeBlocked);
@@ -51,10 +49,37 @@ function Register() {
     ReConfirm: "",
   });
 
+  let users = JSON.parse(localStorage.getItem("Users"))
+    ? JSON.parse(localStorage.getItem("Users"))
+    : [];
+  let emails = [];
+  users.map((value) => {
+    emails.push(value.email);
+  });
+
   let AddUser = () => {
+    if (cont.fname===""||
+    cont.lname===""||
+    cont.email===""||
+    cont.address===""||
+    cont.contact===""||
+    cont.password===""||
+    cont.ReConfirm=== "") {
+      alert("All field should not be empty")
+      
+    }else if (emails.includes(cont.email)) {
+      alert("Email already exists!!");
+    } else {
       table.push(cont);
       localStorage.setItem("Users", JSON.stringify(table));
+      alert("Registered Successfully!!");
+           nav("/");
+    }
+
+
   };
+
+
 
   const validatePassword = (password, confirmPassword) => {
     if (password.length < 8) {
@@ -86,7 +111,7 @@ function Register() {
       setMessage("*Please Enter Valid Email");
     } else if (!regXp.test(cont.email)) {
       setMessage("*Email is not Valid");
-    }else if (!(cont.email.match(regXp))) {
+    } else if (!cont.email.match(regXp)) {
       setMessage("*Email is Invalid");
     } else {
       setMessage("");
@@ -99,134 +124,149 @@ function Register() {
     setCont({ ...cont, [name]: values });
   };
   return (
-    <Box >
-    <Paper
-    className="paper1"
-      elevation={6}
-      sx={{
-        justifyContent: "center",
-        width: { xs: "100%", md: "50%" },
-        paddingBottom: "50px",
-        marginLeft: { md: "25%" },
-        marginTop: "20px",
-          
-      }}>
-      
-      <Typography variant="h4" sx={{ fontWeight: "bolder", pt: "20px", color:"#ff5722", fontFamily:"monospace",textTransform:"uppercase"}}>
-        Registration form
-      </Typography>
-      <br />
-      <Grid container direction={"column"} display={"flex"} spacing={2}>
-        <Grid item>
-          <TextField
-            variant="outlined"
-            label="First Name"
-            sx={{ width: "60%",background:"" }}
-            value={cont.fname}
-            name="fname"
-            onChange={HandleChange}
-            autoComplete="off"
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            variant="outlined"
-            label="Last Name"
-            sx={{ width: "60%" }}
-            value={cont.lname}
-            name="lname"
-            onChange={HandleChange}
-            autoComplete="off"
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            variant="outlined"
-            label="Email"
-            sx={{ width: "60%" }}
-            value={cont.email}
-            name="email"
-            onChange={HandleChange}
-            autoComplete="off"
-            onKeyUp={EmailValidation}
-            helperText={<span style={{ color: "red" }}>{message}</span>}
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            variant="outlined"
-            label="Address"
-            multiline
-            maxRows={3}
-            sx={{ width: "60%" }}
-            value={cont.address}
-            name="address"
-            onChange={HandleChange}
-            autoComplete="off"
-          />
-        </Grid>
-        <Grid item>
-          <TextField
-            variant="outlined"
-            label="Contact"
-            sx={{ width: "60%" }}
-            value={cont.contact}
-            name="contact"
-            onChange={HandleChange}
-            autoComplete="off"
-          />
-        </Grid>
-        <Grid item display={"flex"} justifyContent="center" ml={5}>
-          <TextField
-            variant="outlined"
-            label="Password"
-            type={types}
-            sx={{ width: "60%" }}
-            value={cont.password}
-            name="password"
-            onChange={HandleChange}
-            autoComplete="off"
-            helperText={<span style={{ color: "red" }}>{error}</span>}
-            onKeyDown={handleError}
-          />
-          <Button onClick={HandleToggle}>
-            <Icon icon={icon} size={25}></Icon>
-          </Button>
-        </Grid>
-        <Grid item display={"flex"} justifyContent="center" ml={5}>
-          <TextField
-            variant="outlined"
-            label="Confirm Password"
-            type={types1}
-            sx={{ width: "60%" }}
-            value={cont.ReConfirm}
-            name="ReConfirm"
-            onChange={HandleChange}
-            autoComplete="off"
-            helperText={<span style={{ color: "red" }}>{error}</span>}
-            onKeyUp={handleError}
-          />
-          <Button onClick={HandleToggle1}>
-            <Icon icon={icon1} size={25}></Icon>
-          </Button>
-        </Grid>
-        <Grid item display={"flex"} justifyContent="space-evenly">
-          <Button onClick={() => {AddUser();nav("/")}} variant="contained" sx={{backgroundColor:"#3f51b5"}}>
-            Register
-          </Button>
-          <Button variant="contained" color="error">
-            Cancel
-          </Button>
-          
-        </Grid>
-        <br />
-        <Typography>
-        Already Have An Account? <Link to={"/"}>Login</Link>
+    <Box>
+      <Paper
+        className="paper1"
+        elevation={6}
+        sx={{
+          justifyContent: "center",
+          width: { xs: "100%", md: "50%" },
+          paddingBottom: "50px",
+          marginLeft: { md: "25%" },
+          marginTop: "20px",
+        }}>
+        <Typography
+          variant="h4"
+          sx={{
+            fontWeight: "bolder",
+            pt: "20px",
+            color: "#ff5722",
+            fontFamily: "monospace",
+            textTransform: "uppercase",
+          }}>
+          Registration form
         </Typography>
-        <Grid >
+        <br />
+        <Grid container direction={"column"} display={"flex"} spacing={2}>
+          <Grid item>
+            <TextField
+              variant="outlined"
+              label="First Name"
+              sx={{ width: "60%", background: "" }}
+              value={cont.fname}
+              name="fname"
+              onChange={HandleChange}
+              autoComplete="off"
+            />
+          </Grid>
+          <Grid item>
+            <TextField
+              variant="outlined"
+              label="Last Name"
+              sx={{ width: "60%" }}
+              value={cont.lname}
+              name="lname"
+              onChange={HandleChange}
+              autoComplete="off"
+            />
+          </Grid>
+          <Grid item>
+            <TextField
+              variant="outlined"
+              label="Email"
+              sx={{ width: "60%" }}
+              value={cont.email}
+              name="email"
+              onChange={HandleChange}
+              autoComplete="off"
+              onKeyUp={EmailValidation}
+              helperText={<span style={{ color: "red" }}>{message}</span>}
+            />
+          </Grid>
+          <Grid item>
+            <TextField
+              variant="outlined"
+              label="Address"
+              multiline
+              maxRows={3}
+              sx={{ width: "60%" }}
+              value={cont.address}
+              name="address"
+              onChange={HandleChange}
+              autoComplete="off"
+            />
+          </Grid>
+          <Grid item>
+            <TextField
+              variant="outlined"
+              label="Contact"
+              sx={{ width: "60%" }}
+              value={cont.contact}
+              name="contact"
+              onChange={HandleChange}
+              autoComplete="off"
+            />
+          </Grid>
+          <Grid item display={"flex"} justifyContent="center" ml={5}>
+            <TextField
+              variant="outlined"
+              label="Password"
+              type={types}
+              sx={{ width: "60%" }}
+              value={cont.password}
+              name="password"
+              onChange={HandleChange}
+              autoComplete="off"
+              helperText={<span style={{ color: "red" }}>{error}</span>}
+              onKeyDown={handleError}
+            />
+            <Button onClick={HandleToggle}>
+              <Icon icon={icon} size={25}></Icon>
+            </Button>
+          </Grid>
+          <Grid item display={"flex"} justifyContent="center" ml={5}>
+            <TextField
+              variant="outlined"
+              label="Confirm Password"
+              type={types1}
+              sx={{ width: "60%" }}
+              value={cont.ReConfirm}
+              name="ReConfirm"
+              onChange={HandleChange}
+              autoComplete="off"
+              helperText={<span style={{ color: "red" }}>{error}</span>}
+              onKeyUp={handleError}
+            />
+            <Button onClick={HandleToggle1}>
+              <Icon icon={icon1} size={25}></Icon>
+            </Button>
+          </Grid>
+          <Grid item display={"flex"} justifyContent="space-evenly">
+            <Button
+              onClick={() => {
+                AddUser();
+               
+              }}
+              variant="contained"
+              sx={{ backgroundColor: "#3f51b5" }}>
+              Register
+            </Button>
+            <Button
+              variant="contained"
+              onClick={() => {
+                nav("/");
+              }}
+              color="error">
+              Cancel
+            </Button>
+          </Grid>
+          <br />
+          <Typography>
+            Already Have An Account? <Link to={"/"}>Login</Link>
+          </Typography>
+          <Grid></Grid>
         </Grid>
-      </Grid>
-    </Paper>
+      </Paper>
     </Box>
   );
 }
